refactor(portal-base): move configuration module to subdirectory
