+++ target-platforms/packages/portal-base/src/util/configuration.js
